requirements.txt
pillow
pyyaml
streamlit
streamlit-authenticator
google-cloud-aiplatform
google-cloud-logging